adjustments
# Initialization
- offset params changed from uintptr_t to unsigned int
- dereferencing globals is now left for the user to handle (initialization is easier)

# Class Functions
- functions returning pointers are handled better.
- adjusted layout

# Project Example
- numpad 0 should kill the local player
- exit with `END` key

+++ SDK/Basic_Functions.h
@@ -16,7 +16,7 @@ namespace HEXINTON
 	// --------------------------------------------------
 	// # Global functions
 	// --------------------------------------------------
-	bool InitSdk(const std::string& moduleName, uintptr_t gGameMan, uintptr_t gGameDataMan, uintptr_t gWorldCharMan);
+	bool InitSdk(const std::string& moduleName, unsigned int gGameMan, unsigned int gGameDataMan, unsigned int gWorldCharMan);
 	bool InitSdk();
 	void ShutdownSdk();
